modified recipe GET routes to only include relevant association data
